LIBDRUM-380. Move XMLUI local UMD customizations to edu.umd.lib package
Moving XMLUI customizations.
Replaced the DSpace license header with the UMD license header.

https://issues.umd.edu/browse/LIBDRUM-380
