+++ dspace/modules/xmlui/src/main/java/edu/umd/lib/dspace/app/xmlui/aspect/administrative/etd_departmens/ManageETDDepartmentsMain.java
@@ -1,11 +1,8 @@
-/**
- * The contents of this file are subject to the license and copyright
- * detailed in the LICENSE and NOTICE files at the root of the source
- * tree and available online at
+/*
+ * Copyright (c) 2009 The University of Maryland. All Rights Reserved.
  *
- * http://www.dspace.org/license/
  */
-package org.dspace.app.xmlui.aspect.administrative.etd_departments;
+package edu.umd.lib.dspace.app.xmlui.aspect.administrative.etd_departmens;
 
 import java.sql.SQLException;
 
@@ -201,7 +198,7 @@ public class ManageETDDepartmentsMain extends AbstractDSpaceTransformer
         else
         {
             search.addPara().addButton("submit_delete")
-            .setValue(T_submit_delete);
+                    .setValue(T_submit_delete);
         }
 
         search.addHidden("administrative-continue").setValue(knot.getId());
