+++ dspace/modules/xmlui/src/main/java/edu/umd/lib/dspace/app/xmlui/aspect/administrative/units/DeleteUnitsConfirm.java
@@ -1,11 +1,8 @@
-/**
- * The contents of this file are subject to the license and copyright
- * detailed in the LICENSE and NOTICE files at the root of the source
- * tree and available online at
+/*
+ * Copyright (c) 2009 The University of Maryland. All Rights Reserved.
  *
- * http://www.dspace.org/license/
  */
-package org.dspace.app.xmlui.aspect.administrative.units;
+package edu.umd.lib.dspace.app.xmlui.aspect.administrative.units;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -23,9 +20,9 @@ import org.dspace.authorize.AuthorizeException;
 import org.dspace.eperson.Unit;
 
 /**
- * Present the user with a list of soon-to-be-deleted Groups. If the user clicks
+ * Present the user with a list of soon-to-be-deleted Units. If the user clicks
  * confirm deletion then they will be deleted otherwise they will be spared the
- * wrath of deletion.
+ * wrath of deletion. Similar to DeleteGroupsConfirm
  *
  * @author Scott Phillips
  */
@@ -67,7 +64,7 @@ public class DeleteUnitsConfirm extends AbstractDSpaceTransformer
 
     @Override
     public void addBody(Body body) throws WingException, SQLException,
-            AuthorizeException
+    AuthorizeException
     {
         String idsString = parameters.getParameter("unitIDs", null);
